drivers: thermal: bcl_peripheral: Read charge depletion percentage
soc driver is used to read percentage of charge left on the battery,
which will require the thermal framework to monitor the value falling
below a threshold. This support is not natively available.

soc driver will read the charge depletion percentage, which will
indicate the percentage of charge depleted in the battery. This can use
the native trip threshold monitoring functionality of thermal framework.

[Duckyduck: now it reads batt_psy correctly]
[   84.213876] bcl_peripheral:bcl_write_general_register wrote 0x04 to 0x4248
[   84.242127] bcl_peripheral:bcl_read_soc soc:52
[   84.342606] bcl_peripheral:bcl_read_vbat vbat:3666 mv
[   84.342973] bcl_peripheral:bcl_read_soc soc:52
[   84.347856] bcl_peripheral:bcl_write_general_register wrote 0x04 to 0x4248
[   84.476790] bcl_peripheral:bcl_read_vbat vbat:3666 mv
[   84.477150] bcl_peripheral:bcl_read_soc soc:52
[   84.482321] bcl_peripheral:bcl_write_general_register wrote 0x04 to 0x4248
[   84.610987] bcl_peripheral:bcl_read_soc soc:52
[   84.611244] bcl_peripheral:bcl_read_vbat vbat:3666 mv

Change-Id: I8c332910c08df2573ab564be96b82c642ac7701e
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>